utils/scraped_images/requirements.txt
Pillow
request
google_images_download
